[ADD] New module medical_disease extracted from medical_emr
