+++ __unported__/medical_emr/__openerp__.py
@@ -45,10 +45,8 @@
     "data": [
         'views/medical_sequence.xml',
         'views/medical_secondary_condition_view.xml',
-        'views/medical_pathology_category_view.xml',
         'views/medical_signs_and_symptoms_view.xml',
         'views/medical_directions_view.xml',
-        'views/medical_pathology_view.xml',
         'views/medical_ethnicity_view.xml',
         'views/medical_prescription_order_view.xml',
         'views/medical_medicament_category_view.xml',
@@ -61,13 +59,11 @@
         'views/medical_patient_medication_view.xml',
         'views/medical_patient_evaluation_view.xml',
         'views/medical_prescription_line_view.xml',
-        'views/medical_patient_disease_view.xml',
         'views/medical_patient_view.xml',
         'views/medical_drug_route_view.xml',
         'views/medical_family_view.xml',
         'views/medical_occupation_view.xml',
         'views/medical_medicament_view.xml',
-        'views/medical_pathology_group_view.xml',
         'security/ir.model.access.csv',
         'medical_menu.xml',
     ],
