+++ __unported__/medical_emr/medical_menu.xml
@@ -19,31 +19,6 @@
             parent="medical_prescription_root" sequence="10"
             action="medical_prescription_order_action" />
 
-<!-- Diseases -->
-        <menuitem id="medical_root_disease" name="Disease"
-            parent="medical.medical_root_sub" sequence="10" />
-
-<!--  Menu of Pathology-->
-
-        <menuitem id="medical_sub_disease"
-            parent="medical_root_disease" sequence="1" 
-            action="medical_patient_disease_action" />
-
-        <menuitem id="medical_pathology"
-            parent="medical_root_disease" sequence="10" 
-            action="medical_pathology_action" />
-
-        <menuitem id="medical_pathology_category"
-            parent="medical_root_disease" sequence="20"
-            action="medical_pathology_category_action" />
-
-        <menuitem id="medical_pathology_group"
-            parent="medical_root_disease" sequence="30"
-            action="medical_pathology_group_action" />
-
-        <menuitem id="medical_root_disease_cat" name="Categories"
-            parent="" sequence="100" />
-
 <!-- Occupation  ( Config/misc/Occupation) )--> 
         <menuitem id="medical_occupation" name="Occupation"
             parent="medical.medical_root_misc" sequence="20"
