+++ medical_disease/models/medical_pathology.py
