+++ medical_disease/models/medical_pathology_category.py
