+++ medical_disease/models/medical_pathology_group.py
