+++ medical_disease/security/ir.model.access.csv
@@ -1,0 +1,13 @@
+id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
+access_medical_pathology_group_assistant,medical.pathology_group_assistant,model_medical_pathology_group,medical.group_medical_assistant,1,1,1,0
+access_medical_pathology_assistant,medical.pathology_assistant,model_medical_pathology,medical.group_medical_assistant,1,1,1,0
+access_medical_pathology_category_assistant,medical.pathology_category_assistant,model_medical_pathology_category,medical.group_medical_assistant,1,1,1,0
+access_medical_pathology_group_doctor,medical.pathology_group_doctor,model_medical_pathology_group,medical.group_medical_doctor,1,1,1,0
+access_medical_pathology_doctor,medical.pathology_doctor,model_medical_pathology,medical.group_medical_doctor,1,1,1,0
+access_medical_pathology_category_doctor,medical.pathology_category_doctor,model_medical_pathology_category,medical.group_medical_doctor,1,1,1,0
+access_medical_pathology_group_user,medical.pathology_group_user,model_medical_pathology_group,medical.group_medical_user,1,0,0,0
+access_medical_pathology_user,medical.pathology_user,model_medical_pathology,medical.group_medical_user,1,0,0,0
+access_medical_pathology_category_user,medical.pathology_category_user,model_medical_pathology_category,medical.group_medical_user,1,0,0,0
+access_medical_patient_disease_assistant,medical.patient_disease_assistant,model_medical_patient_disease,medical.group_medical_assistant,1,1,1,0
+access_medical_patient_disease_user,medical.patient_disease_user,model_medical_patient_disease,medical.group_medical_user,1,0,0,0
+access_medical_patient_disease_doctor,medical.patient_disease_doctor,model_medical_patient_disease,medical.group_medical_doctor,1,1,1,0
