+++ medical_disease/views/medical_pathology_category_view.xml
