+++ medical_disease/views/medical_pathology_view.xml
