+++ medical_disease/views/medical_patient_view.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<openerp>
+    <data>
+        <record id="medical_patient_view_form" model="ir.ui.view">
+            <field name="name">medical.patient.form (medical_disease)</field>
+            <field name="model">medical.patient</field>
+            <field name="inherit_id" ref="medical.view_medical_patient_form" />
+            <field name="arch" type="xml">
+                <xpath expr="//div[@name='buttons']" position="inside">
+                    <button
+                        name="%(medical_disease.medical_patient_disease_action)d"
+                        type="action" class="oe_stat_button" icon="fa-list"
+                        context="{'search_default_patient_id': active_id}">
+                        <div>Diseases</div>
+                    </button>
+                </xpath>
+            </field>
+        </record>
+    </data>
+</openerp>
